Refactor ui components to components folder
