+++ Dockerfile
@@ -1,12 +1,15 @@
-# Imagem de Origem
 FROM node:15-alpine
-# Diretório de trabalho(é onde a aplicação ficará dentro do container).
+
 WORKDIR /app
-# Adicionando `/app/node_modules/.bin` para o $PATH
+
 ENV PATH /app/node_modules/.bin:$PATH
-# Instalando dependências da aplicação e armazenando em cache.
-COPY package.json /app/package.json
+
+COPY package.json ./
+COPY package-lock.json ./
+
 RUN npm install --silent
 RUN npm install react-scripts@3.4.4 -g --silent
-# start app
+
+COPY . ./
+
 CMD ["npm", "start"]
